backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.5
blinker==1.9.0
cachelib==0.13.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
deprecation==2.1.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Session==0.8.0
google-auth==2.43.0
google-oauth==1.0.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.19.0
multidict==6.7.0
packaging==25.0
postgrest==2.24.0
pytz==2025.2
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.1.1
realtime==2.24.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
